requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
gspread==6.1.2
google-auth==2.29.0
